requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-cloudinary-storage==0.3.0
django-etc==1.4.0
django-taggit==6.0.0
djangorestframework==3.15.2
idna==3.8
oauthlib==3.2.2
pillow==10.4.0
pip-autoremove==0.10.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.2
